Rename classloaders module
