--- instrumentation/classloaders/javaagent-integration-tests/javaagent-integration-tests.gradle
@@ -1,8 +1,0 @@
-ext.skipPublish = true
-
-apply from: "$rootDir/gradle/instrumentation.gradle"
-
-dependencies {
-  testImplementation group: "org.apache.commons", name: "commons-lang3", version: "3.12.0"
-  testImplementation group: "commons-io", name: "commons-io", version: "2.8.0"
-}
