--- instrumentation/classloaders/javaagent-unittests/classloaders-javaagent-unittests.gradle
@@ -1,6 +1,0 @@
-apply from: "$rootDir/gradle/java.gradle"
-
-dependencies {
-  testImplementation project(':instrumentation:classloaders:javaagent')
-  testImplementation project(':javaagent-bootstrap')
-}
