+++ instrumentation/internal-class-loader/javaagent/src/test/groovy/ClassLoadingTest.groovy
