+++ instrumentation/internal-class-loader/javaagent/src/test/groovy/JBossClassloadingTest.groovy
