+++ instrumentation/internal-class-loader/javaagent/src/test/groovy/OSGIClassloadingTest.groovy
