+++ instrumentation/internal-class-loader/javaagent/src/test/groovy/ResourceInjectionTest.groovy
