+++ instrumentation/internal-class-loader/javaagent/src/test/resources/test-resources/test-resource.txt
