+++ instrumentation/internal-proxy/javaagent-unittests/internal-proxy-javaagent-unittests.gradle
@@ -1,0 +1,6 @@
+apply from: "$rootDir/gradle/java.gradle"
+
+dependencies {
+  testImplementation project(':instrumentation:internal-proxy:javaagent')
+  testImplementation project(':javaagent-bootstrap')
+}
