+++ instrumentation/internal-proxy/javaagent/src/test/groovy/NewProxyInstanceTest.groovy
