+++ instrumentation/internal-url-class-loader/javaagent-integration-tests/internal-url-class-loader-javaagent-integration-tests.gradle
@@ -1,0 +1,10 @@
+ext.skipPublish = true
+
+apply from: "$rootDir/gradle/instrumentation.gradle"
+
+dependencies {
+  testImplementation group: "org.apache.commons", name: "commons-lang3", version: "3.12.0"
+  testImplementation group: "commons-io", name: "commons-io", version: "2.8.0"
+
+  testInstrumentation project(":instrumentation:internal-url-class-loader:javaagent")
+}
