+++ instrumentation/internal-url-class-loader/javaagent-integration-tests/src/test/groovy/AddUrlTest.groovy
