+++ testing/agent-exporter/agent-exporter.gradle
@@ -24,7 +24,7 @@ dependencies {
   implementation project(':instrumentation:executors:javaagent')
   // FIXME: we should enable this, but currently this fails tests for google http client
   //testImplementation project(':instrumentation:http-url-connection:javaagent')
-  implementation project(':instrumentation:classloaders:javaagent')
+  implementation project(':instrumentation:internal-class-loader:javaagent')
 
   // Many tests use OpenTelemetry API calls, e.g., via TraceUtils.runUnderTrace
   implementation project(':instrumentation:opentelemetry-annotations-1.0:javaagent')
